refactor: split alarm-related routes into separate alarms module for improved organization
